+++ Classes/Configuration/Configurator.php
@@ -51,7 +51,7 @@ class Configurator
             if ($rootPageForTsConfigRow !== null) {
                 $rootPageForTsConfig = $rootPageForTsConfigRow['uid'];
             }
-            $serviceConfig = GeneralUtility::makeInstance('TYPO3\\CMS\\Extbase\\Service\\TypoScriptService')
+            $serviceConfig = GeneralUtility::makeInstance(\TYPO3\CMS\Core\TypoScript\TypoScriptService::class)
                 ->convertTypoScriptArrayToPlainArray(BackendUtility::getPagesTSconfig($rootPageForTsConfig));
             if (isset($serviceConfig['tx_translatr'])) {
                 $this->setConfig($serviceConfig['tx_translatr']);