+++ ext_emconf.php
@@ -14,7 +14,7 @@ $EM_CONF[$_EXTKEY] = [
     'version' => '1.0.0',
     'constraints' => [
         'depends' => [
-            'typo3' => '8.7.0-9.5.999',
+            'typo3' => '9.5.999-10.4.999',
             'vhs' => '5.0.0-6.99.999',
         ],
         'conflicts' => [